webpack-reporter: update to archetype 2 with src/lib mode (#259)
* webpack-reporter: update to archetype 2 with src/lib mode

* webpack-reporter: fix tests
