--- packages/electrode-webpack-reporter/.babelrc
@@ -1,3 +1,0 @@
-{
-  "extends": "./node_modules/electrode-archetype-react-app/config/babel/.babelrc"
-}
