+++ packages/electrode-webpack-reporter/config/default.json
@@ -9,13 +9,13 @@
         "pathPrefix": "dist"
       }
     },
-    "webapp": {
-      "module": "./server/plugins/webapp",
+    "electrode-react-webapp/lib/hapi": {
       "options": {
+        "serverSideRendering": false,
         "paths": {
           "/{args*}": {
             "content": {
-              "module": "./server/views/index-view"
+              "module": "./{{env.APP_SRC_DIR}}/server/views/index-view"
             }
           }
         }
