+++ packages/electrode-webpack-reporter/src/client/.babelrc
@@ -1,0 +1,3 @@
+{
+  "extends": "electrode-archetype-react-app-dev/config/babel/babelrc-client"
+}
