+++ packages/electrode-webpack-reporter/src/client/components/home.jsx
