+++ packages/electrode-webpack-reporter/src/client/components/legacy.jsx
