+++ packages/electrode-webpack-reporter/src/client/components/navbar.jsx
