+++ packages/electrode-webpack-reporter/src/client/components/report.jsx
