+++ packages/electrode-webpack-reporter/src/client/index.html
