+++ packages/electrode-webpack-reporter/src/client/reducers/index.jsx
