+++ packages/electrode-webpack-reporter/src/server/.babelrc
@@ -1,0 +1,3 @@
+{
+  "extends": "electrode-archetype-react-app-dev/config/babel/babelrc-server"
+}
