+++ packages/electrode-webpack-reporter/src/server/README.md
@@ -1,0 +1,5 @@
+This directory is only use for development of the Reporter.
+
+When publishing, only the reporter's browser bundle is published.
+
+The reporter plugin will add a route to webpack-dev-server's express server and serve the bundle to be executed on the browser client side.
