+++ packages/electrode-webpack-reporter/src/server/make-init-state.js
