+++ packages/electrode-webpack-reporter/src/server/plugins/webapp/index.html
