+++ packages/electrode-webpack-reporter/src/server/views/index-view.jsx
