+++ packages/electrode-webpack-reporter/test/client/modules.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 module.exports = {
     "chunks":[
       {
